% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR_95CI.R
\name{OR_95CI}
\alias{OR_95CI}
\title{Odds ratios and 95 percent confidence intervals}
\usage{
OR_95CI(coef, se, siglevel, roundto)
}
\arguments{
\item{coef}{A numeric vector of coefficient estimates (log-odds) from a statistical model.}

\item{se}{A numeric vector of standard errors corresponding to the coefficient estimates.}

\item{siglevel}{Significance level (e.g., 0.05 for a 95 percent confidence interval).}

\item{roundto}{Integer indicating the number of decimal places to round the results.}
}
\value{
A character vector where each element contains the odds ratio along with its confidence interval in the format: "OR (lower CI, upper CI)".
}
\description{
Returns odds ratios and 95 percent confidence intervals from a vector of coeﬀicient estimates and a vector of standard errors.
}
\examples{
# load toydata
data(toydata)

# fit a logistic regression model
mod <- glm(y ~ x1 + x2, data = toydata, family = binomial)

# extract coefficients and standard errors
coef <- coef(mod)
se <- sqrt(diag(vcov(mod)))

# calculate odds ratios and 95\% confidence intervals
OR_95CI(coef, se, siglevel = 0.05, roundto = 2)
}
